requirements.txt
torch
torchvision
pandas
numpy
matplotlib
scikit-learn
pyyaml
yfinance
tqdm
